backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
pydantic>=2.9.0

# LangChain + LangGraph
langgraph>=0.2.0
langgraph-prebuilt>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
langchain-groq>=0.2.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Document processing
pypdf>=5.0.0
python-multipart>=0.0.12

# Utilities
uuid6>=2024.1.12
